refactor: Change `spent_time` unit from minutes to decimal hours across schema, model, and views.

+++ plugins/redmine_operational_metrics/db/migrate/001_create_operational_metrics.rb
@@ -6,7 +6,7 @@ class CreateOperationalMetrics < ActiveRecord::Migration[5.2]
       t.string :project
       t.text :task_planed
       t.date :task_date
-      t.integer :spent_time
+      t.decimal :spent_time
       t.text :spent_detail
       t.string :collaborate_with
       t.string :completion, null: false
